Moving constraint logic to annotations (#761)
