+++ calm-hub/src/test/java/org/finos/calm/resources/TestNamespaceResourceShould.java
@@ -78,7 +78,7 @@ public class TestNamespaceResourceShould {
                 .post("/calm/namespaces")
                 .then()
                 .statusCode(400)
-                .body(containsString("Namespace name is required"));
+                .body(containsString("Namespace must not be null"));
 
         verify(namespaceStore, never()).namespaceExists(any());
         verify(namespaceStore, never()).createNamespace(any());
@@ -93,7 +93,7 @@ public class TestNamespaceResourceShould {
                 .post("/calm/namespaces")
                 .then()
                 .statusCode(400)
-                .body(containsString("Namespace name is required"));
+                .body(containsString("Namespace must not be blank"));
 
         verify(namespaceStore, never()).namespaceExists(any());
         verify(namespaceStore, never()).createNamespace(any());
@@ -108,7 +108,7 @@ public class TestNamespaceResourceShould {
                 .post("/calm/namespaces")
                 .then()
                 .statusCode(400)
-                .body(containsString("Namespace must contain only alphanumeric characters and hyphens"));
+                .body(containsString("Namespace must match pattern: ^[A-Za-z0-9-]+$"));
 
         verify(namespaceStore, never()).namespaceExists(any());
         verify(namespaceStore, never()).createNamespace(any());
@@ -139,26 +139,9 @@ public class TestNamespaceResourceShould {
                 .post("/calm/namespaces")
                 .then()
                 .statusCode(400)
-                .body(containsString("Namespace name is required"));
+                .body(containsString("Request must not be null"));
 
         verify(namespaceStore, never()).namespaceExists(any());
         verify(namespaceStore, never()).createNamespace(any());
-    }
-
-    @Test
-    void trim_whitespace_from_namespace_name() {
-        when(namespaceStore.namespaceExists("trimmed-namespace")).thenReturn(false);
-
-        given()
-                .contentType("application/json")
-                .body("{\"namespace\":\"  trimmed-namespace  \"}")
-                .when()
-                .post("/calm/namespaces")
-                .then()
-                .statusCode(201)
-                .header("Location", containsString("/calm/namespaces/trimmed-namespace"));
-
-        verify(namespaceStore).namespaceExists("trimmed-namespace");
-        verify(namespaceStore).createNamespace("trimmed-namespace");
     }
 }
